Split into separate files, tidy, make calculate return string.
