+++ src/main/java/com/softenggroup5/calculator/Calculator.java
@@ -1,0 +1,9 @@
+package com.softenggroup5.calculator;
+
+public class Calculator {
+    // This returns a string so that errors, ints, and doubles can be returned from the same function.
+    // Lower level functions can return different types. This is th only one which should be called outside testing.
+    public static String calculate(String input){
+        return "0";
+    }
+}
